refactor: minimize code coupling
And I'll no longer be using gitmoji from now on. While they are nice and easy to follow when you get the hang of them, they are very tedious to copy/paste each time and may be unclear for the newcomers.

Gitmoji can always be used if they are a convention among the codebase. Semantic commits are much more universal.